wpa: reduce DETACH timeout to 10ms
There's really no need to wait for DETACH to complete. We do not only for
compatibility and to make sure the message is actually transmitted.
However, if wpa-supplicant is currently terminating, the message might be
in the queue, but not processed. We thus don't get any ECONNREFUSED and
wait for 10s for an answer.. That's just awful.

Set it to 10ms for now as that seems to work just fine.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>