wpa: reduce default timeout to 1s
Instead of waiting 10s for an answer, lets use 1s as default.
wpa-supplicant sometimes performs blocking actions that *might* take a
while. However, waiting 10s just screws the whole applications and is the
wrong approach to fix that. wpa_supplicant needs to learn asynchronous
transactions and deal with incoming requests in a timely manner.
Therefore, set the default to 1s and be happy \o/

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>